bugfix(annotaterb): Annotate pre-commit hook never worked, so removed
Seems that there's no annotaterb command that will exhibit the behavior
we want here. That's to say: check to see if files need to be annotated,
and exit non-zero if they do, or zero if they don't, but don't change
anything.

Until something like that is in place, it's best to remove it from the
hooks and just depend on it being run in ./bin/setup